Add basic PageXML export
- PageXMLDoc classes now have a `to_pagexml` method to export the
  content to a PageXML XML tree.
- Documment model classes and functionality has been refactored in
  preparation for a API overhaul.
